+++ tests/tests/rest_adapter.js
@@ -63,9 +63,8 @@ test('creates valid path for multi-word model classes', function() {
 });
 
 asyncTest('can optionally add query params to a findByKey request', 1, function() {
-  var person = App.Person.find({ id: 1, some_param: 'test' });
-
-  person.get('currentRequest').always(function() {
+  var personFetch = App.Person.fetch({ id: 1, some_param: 'test' });
+  personFetch._currentRequest.abort().always(function() {
     var urlParts = this.url.split('/');
     var path = urlParts[urlParts.length-1];
     equal( path, '1?some_param=test', 'findByKey with parameters requests expected url' );